Use bash instead of cert_cp.py
more subcommands

+++ cert-up.sh
@@ -1,17 +1,26 @@
 #!/bin/bash
 
+CERT_FILES=(
+  'cert.pem'
+  'privkey.pem'
+  'fullchain.pem'
+)
+
 # path of this script
-BASE_ROOT=$(cd "$(dirname "$0")";pwd)
+BASE_ROOT=$(
+  cd "$(dirname "$0")"
+  pwd
+)
 # date time
-DATE_TIME=`date +%Y%m%d%H%M%S`
+DATE_TIME=$(date +%Y%m%d%H%M%S)
 # base crt path
 CRT_BASE_PATH=/usr/syno/etc/certificate
 PKG_CRT_BASE_PATH=/usr/local/etc/certificate
 ACME_BIN_PATH=~/.acme.sh
 ACME_CRT_PATH=~/certificates
 TEMP_PATH=~/temp
-CRT_PATH_NAME=`sudo cat ${CRT_BASE_PATH}/_archive/DEFAULT`
-CRT_PATH=${CRT_BASE_PATH}/_archive/${CRT_PATH_NAME}
+ARCHIEV_PATH="${CRT_BASE_PATH}/_archive"
+INFO_FILE_PATH="${ARCHIEV_PATH}/INFO"
 
 SYNCTHING_PATH=/volume1/@appdata/syncthing
 # Before DSM 7.0, use the following instead:
@@ -38,18 +47,18 @@ if [ "$has_config" = true ] && [ ! -f ${ACME_BIN_PATH}/registed ]; then
   do_register=true
 fi
 
-backup_cert () {
+backup_cert() {
   echo 'begin backup_cert'
   BACKUP_PATH=~/cert_backup/${DATE_TIME}
   mkdir -p ${BACKUP_PATH}
   sudo cp -r ${CRT_BASE_PATH} ${BACKUP_PATH}
   sudo cp -r ${PKG_CRT_BASE_PATH} ${BACKUP_PATH}/package_cert
-  echo ${BACKUP_PATH} > ~/cert_backup/latest
+  echo ${BACKUP_PATH} >~/cert_backup/latest
   echo 'done backup_cert'
   return 0
 }
 
-install_acme () {
+install_acme() {
   echo 'begin install_acme'
   mkdir -p ${TEMP_PATH}
   rm -rf ${TEMP_PATH}/acme.sh
@@ -61,13 +70,13 @@ install_acme () {
   curl -L -o ${SRC_TAR_NAME} ${ACME_SH_ADDRESS}
   if [ ! -f ${TEMP_PATH}/${SRC_TAR_NAME} ]; then
     echo 'download failed'
-    exit 1;
+    exit 1
   fi
-  SRC_NAME=`tar -tzf ${SRC_TAR_NAME} | head -1 | cut -f1 -d"/"`
+  SRC_NAME=$(tar -tzf ${SRC_TAR_NAME} | head -1 | cut -f1 -d"/")
   tar zxvf ${SRC_TAR_NAME}
   if [ ! -d ${TEMP_PATH}/${SRC_NAME} ]; then
     echo 'the file downloaded incorrectly'
-    exit 1;
+    exit 1
   fi
   # OR git master
   #git clone https://github.com/acmesh-official/acme.sh.git
@@ -91,19 +100,19 @@ install_acme () {
   return 0
 }
 
-register_account () {
+register_account() {
   cd ${ACME_BIN_PATH}
   ${ACME_BIN_PATH}/acme.sh --register-account -m "${EMAIL}"
   if [ $? -ne 0 ]; then
     echo 'register_account failed!!'
-    return 1;
+    return 1
   fi
   touch ${ACME_BIN_PATH}/registed
   echo 'done register_account'
   return 0
 }
 
-generate_cert () {
+generate_cert() {
   if [ "$do_register" = true ]; then
     # first register account
     register_account
@@ -113,44 +122,82 @@ generate_cert () {
   fi
   echo 'begin generate_cert'
   cd ${ACME_BIN_PATH}
-  echo 'begin updating default cert by acme.sh tool'
   source "${ACME_BIN_PATH}/acme.sh.env"
-  ${ACME_BIN_PATH}/acme.sh --force --log --issue --dns ${DNS} --dnssleep ${DNS_SLEEP} -d \"${DOMAIN}\" -d \"*.${DOMAIN}\"
+  echo 'begin updating default cert by acme.sh tool'
+  ${ACME_BIN_PATH}/acme.sh --force --log --issue --dns ${DNS} --dnssleep ${DNS_SLEEP} -d "${DOMAIN}" -d "*.${DOMAIN}"
   #   --cert-file ${ACME_CRT_PATH}/cert.pem \
   #   --key-file ${ACME_CRT_PATH}/privkey.pem \
   #   --ca-file ${ACME_CRT_PATH}/chain.pem \
   #   --fullchain-file ${ACME_CRT_PATH}/fullchain.pem
   if [ $? -eq 0 ] && [ -s ${ACME_CRT_PATH}/${DOMAIN}/ca.cer ]; then
-    sudo cp ${ACME_CRT_PATH}/${DOMAIN}/ca.cer ${CRT_PATH}/chain.pem
-    sudo cp ${ACME_CRT_PATH}/${DOMAIN}/fullchain.cer ${CRT_PATH}/fullchain.pem
-    sudo cp ${ACME_CRT_PATH}/${DOMAIN}/${DOMAIN}.cer ${CRT_PATH}/cert.pem
-    sudo cp ${ACME_CRT_PATH}/${DOMAIN}/${DOMAIN}.key ${CRT_PATH}/privkey.pem
     echo 'done generate_cert'
     return 0
   else
     echo '[ERR] fail to generate_cert'
-    echo "begin revert"
-    revert_cert
-    exit 1;
+    exit 1
   fi
 }
 
-update_service () {
+update_service() {
   echo 'begin update_service'
-  sudo /bin/python ${BASE_ROOT}/cert_cp.py ${CRT_PATH_NAME}
-  if [ "$SYNCTHING" = true ]; then
-    echo 'Copy cert for Syncthing'
-    sudo cp -v --preserve=timestamps ${CRT_PATH}/cert.pem ${SYNCTHING_PATH}/https-cert.pem
-    sudo cp -v --preserve=timestamps ${CRT_PATH}/privkey.pem ${SYNCTHING_PATH}/https-key.pem
-    sudo chmod 664 ${SYNCTHING_PATH}/https-cert.pem
-    sudo chmod 600 ${SYNCTHING_PATH}/https-key.pem
-    sudo chown ${SYNCTHING_USER}:${SYNCTHING_GROUP} ${SYNCTHING_PATH}/https-cert.pem
-    sudo chown ${SYNCTHING_USER}:${SYNCTHING_GROUP} ${SYNCTHING_PATH}/https-key.pem
+
+  CRT_PATH_NAME=$(sudo cat ${ARCHIEV_PATH}/DEFAULT)
+  CRT_PATH=${ARCHIEV_PATH}/${CRT_PATH_NAME}
+  services=()
+  info=$(sudo cat "$INFO_FILE_PATH")
+  if [ -z "$info" ]; then
+    echo "Failed to read file: $INFO_FILE_PATH"
+    exit 1
+  else
+    services=($(echo "$info" | jq -r ".$CRT_PATH_NAME.services[] | @base64"))
   fi
-  echo 'done update_service'
+  if [[ ${#services[@]} -eq 0 ]]; then
+    echo "[ERR] load INFO file - $INFO_FILE_PATH fail"
+    exit 1
+  fi
+
+  if [ -e ${ACME_CRT_PATH}/${DOMAIN}/ca.cer ] && [ -s ${ACME_CRT_PATH}/${DOMAIN}/ca.cer ]; then
+    sudo cp -v ${ACME_CRT_PATH}/${DOMAIN}/ca.cer ${CRT_PATH}/chain.pem
+    sudo cp -v ${ACME_CRT_PATH}/${DOMAIN}/fullchain.cer ${CRT_PATH}/fullchain.pem
+    sudo cp -v ${ACME_CRT_PATH}/${DOMAIN}/${DOMAIN}.cer ${CRT_PATH}/cert.pem
+    sudo cp -v ${ACME_CRT_PATH}/${DOMAIN}/${DOMAIN}.key ${CRT_PATH}/privkey.pem
+    for service in "${services[@]}"; do
+      display_name=$(echo "$service" | base64 --decode | jq -r '.display_name')
+      isPkg=$(echo "$service" | base64 --decode | jq -r '.isPkg')
+      subscriber=$(echo "$service" | base64 --decode | jq -r '.subscriber')
+      service_name=$(echo "$service" | base64 --decode | jq -r '.service')
+
+      echo "Copy cert for $display_name"
+      if [[ $isPkg == true ]]; then
+        CP_TO_DIR="${PKG_CRT_BASE_PATH}/${subscriber}/${service_name}"
+      else
+        CP_TO_DIR="${CRT_BASE_PATH}/${subscriber}/${service_name}"
+      fi
+      for f in "${CERT_FILES[@]}"; do
+        src="$CRT_PATH/$f"
+        des="$CP_TO_DIR/$f"
+        if [[ -e "$des" ]]; then
+          sudo rm -fr "$des"
+        fi
+        sudo cp -v "$src" "$des" || echo "[WRN] copy from $src to $des fail"
+      done
+    done
+    if [ "$SYNCTHING" = true ]; then
+      echo 'Copy cert for Syncthing'
+      sudo cp -v --preserve=timestamps ${CRT_PATH}/cert.pem ${SYNCTHING_PATH}/https-cert.pem
+      sudo cp -v --preserve=timestamps ${CRT_PATH}/privkey.pem ${SYNCTHING_PATH}/https-key.pem
+      sudo chmod 664 ${SYNCTHING_PATH}/https-cert.pem
+      sudo chmod 600 ${SYNCTHING_PATH}/https-key.pem
+      sudo chown ${SYNCTHING_USER}:${SYNCTHING_GROUP} ${SYNCTHING_PATH}/https-cert.pem
+      sudo chown ${SYNCTHING_USER}:${SYNCTHING_GROUP} ${SYNCTHING_PATH}/https-key.pem
+    fi
+    echo 'done update_service'
+  else
+    echo "no cert files, pls run: $0 generate_cert"
+  fi
 }
 
-reload_webservice () {
+reload_webservice() {
   echo 'begin reload_webservice'
 
   # Not all Synology NAS webservice are the same.
@@ -168,11 +215,11 @@ reload_webservice () {
   fi
 }
 
-revert_cert () {
+revert_cert() {
   echo 'begin revert_cert'
   BACKUP_PATH=${BASE_ROOT}/backup/$1
   if [ -z "$1" ]; then
-    BACKUP_PATH=`cat ${BASE_ROOT}/backup/latest`
+    BACKUP_PATH=$(cat ${BASE_ROOT}/backup/latest)
   fi
   if [ ! -d "${BACKUP_PATH}" ]; then
     echo "[ERR] backup path: ${BACKUP_PATH} not found."
@@ -186,37 +233,54 @@ revert_cert () {
   echo 'done revert_cert'
 }
 
-update_cert () {
-  echo '------ begin update_cert ------'
+case "$1" in
+install)
+  echo "------ install script ------"
+  install_acme
+  ;;
+
+backup_cert)
+  echo "------ backup_cert ------"
+  backup_cert
+  ;;
+
+generate_cert)
+  echo "------ generate_cert ------"
+  generate_cert
+  ;;
+
+update_service)
+  echo "------ update_service ------"
+  update_service
+  ;;
+
+reload_webservice)
+  echo "------ reload_webservice ------"
+  reload_webservice
+  ;;
+
+update)
+  echo '------ update_cert ------'
   backup_cert
   generate_cert
   update_service
   reload_webservice
-  echo '------ end update_cert ------'
-}
-
-case "$1" in
-install)
-  echo "begin install script"
-  install_acme
-  ;;
-
-update)
-  echo "begin update cert"
-  update_cert
   ;;
 
 register)
-  echo "begin register_account"
+  echo "------ register_account ------"
   register_account
   ;;
 
 revert)
-  echo "begin revert"
+  echo "------ revert ------"
   revert_cert $2
   ;;
 
 *)
   echo "Usage: $0 {install|update|register|revert}"
+  echo or
+  echo "Usage: $0 {install|backup_cert|generate_cert|update_service|reload_webservice|register|revert}"
   exit 1
+  ;;
 esac
